:tada: :sparkles: Initial commit of SkinvestigatorAI project
This commit includes the core implementation of the SkinvestigatorAI project, with data scraper, model training, and web application components. The project is structured using Pyramid, TensorFlow, and Keras libraries for an efficient and user-friendly skin cancer detection platform. The commit does not include tests, which will be added in subsequent commits. :rocket: :whale: :green_heart: